Add a non-blocking API as well as an example
The non-blocking API can be used by applications that need to process input from
multiple sources and perform work also while waiting for user to finish writing
a command.

Note that this non-blocking API currently only works for real and supported
terminals. Note also that when hiding and showing the prompt we currently choose
to enable and disable raw mode. This is actually not needed, it would be enough
just to just reconfigure the output.